+++ src/common/WiFiSettings/WiFiForm.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2022 CZ.NIC z.s.p.o. (https://www.nic.cz/)
+ * Copyright (C) 2019-2026 CZ.NIC z.s.p.o. (https://www.nic.cz/)
  *
  * This is free software, licensed under the GNU General Public License v3.
  * See /LICENSE for more information.
@@ -10,7 +10,9 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import { HELP_TEXTS, HTMODES, BANDS } from "./constants";
-import getEncryptionChoices from "./encryptionHelpers";
+import getEncryptionChoices, {
+    getSupportedEncryption,
+} from "./encryptionHelpers";
 import WifiGuestForm from "./WiFiGuestForm";
 import WiFiQRCode from "./WiFiQRCode";
 import PasswordInput from "../../bootstrap/PasswordInput";
@@ -171,19 +173,27 @@ function DeviceForm({
                             // Get the last item in the available HT modes for the selected band
                             const bestHtmode =
                                 selectedBand.available_htmodes.slice(-1)[0];
+                            // The band limits the encryption, both for the
+                            // main and for the guest network
                             const deviceUpdate = {
                                 band: { $set: value },
                                 channel: { $set: "0" },
                                 htmode: { $set: bestHtmode },
+                                encryption: {
+                                    $set: getSupportedEncryption(
+                                        value,
+                                        formData.encryption
+                                    ),
+                                },
+                                guest_wifi: {
+                                    encryption: {
+                                        $set: getSupportedEncryption(
+                                            value,
+                                            formData.guest_wifi.encryption
+                                        ),
+                                    },
+                                },
                             };
-                            // The 6 GHz band supports only WPA3, both for the
-                            // main and for the guest network
-                            if (value === "6g") {
-                                deviceUpdate.encryption = { $set: "WPA3" };
-                                deviceUpdate.guest_wifi = {
-                                    encryption: { $set: "WPA3" },
-                                };
-                            }
                             return {
                                 devices: {
                                     [deviceIndex]: deviceUpdate,